+++ src/sys/windows/udp.rs
@@ -18,11 +18,10 @@ use wio::net::SocketAddrBuf;
 use wio::net::UdpSocketExt as WioUdpSocketExt;
 
 use {Evented, EventSet, IpAddr, PollOpt, Selector, Token};
-use bytes::{Buf, MutBuf};
 use event::IoEvent;
 use sys::windows::selector::{Overlapped, Registration};
 use sys::windows::from_raw_arc::FromRawArc;
-use sys::windows::{bad_state, wouldblock, Family};
+use sys::windows::{bad_state, Family};
 
 pub struct UdpSocket {
     imp: Imp,
@@ -112,31 +111,23 @@ impl UdpSocket {
         })
     }
 
-    pub fn send_to<B: Buf>(&self, buf: &mut B, target: &SocketAddr)
-                           -> io::Result<Option<()>> {
-        match self._send_to(buf.bytes(), target) {
-            Ok(n) => { buf.advance(n); Ok(Some(())) }
-            Err(ref e) if e.kind() == io::ErrorKind::WouldBlock => Ok(None),
-            Err(e) => Err(e),
-        }
-    }
-
     /// Note that unlike `TcpStream::write` this function will not attempt to
     /// continue writing `buf` until its entirely written.
     ///
     /// TODO: This... may be wrong in the long run. We're reporting that we
     ///       successfully wrote all of the bytes in `buf` but it's possible
     ///       that we don't actually end up writing all of them!
-    fn _send_to(&self, buf: &[u8], target: &SocketAddr) -> io::Result<usize> {
+    pub fn send_to(&self, buf: &[u8], target: &SocketAddr)
+                   -> io::Result<Option<usize>> {
         let mut me = self.inner();
         let me = &mut *me;
         match me.write {
             State::Empty => {}
-            _ => return Err(wouldblock())
+            _ => return Ok(None),
         }
         let s = try!(me.socket.socket());
         if me.iocp.port().is_none() {
-            return Err(wouldblock())
+            return Ok(None)
         }
         let mut owned_buf = me.iocp.get_buffer(64 * 1024);
         let amt = try!(owned_buf.write(buf));
@@ -147,11 +138,11 @@ impl UdpSocket {
         });
         me.write = State::Pending(owned_buf);
         mem::forget(self.imp.clone());
-        Ok(amt)
+        Ok(Some(amt))
     }
 
-    pub fn recv_from<B: MutBuf>(&self, buf: &mut B)
-                                -> io::Result<Option<SocketAddr>> {
+    pub fn recv_from(&self, mut buf: &mut [u8])
+                     -> io::Result<Option<(usize, SocketAddr)>> {
         let mut me = self.inner();
         match mem::replace(&mut me.read, State::Empty) {
             State::Empty => Ok(None),
@@ -159,13 +150,13 @@ impl UdpSocket {
             State::Ready(data) => {
                 // If we weren't provided enough space to receive the message
                 // then don't actually read any data, just return an error.
-                if buf.remaining() < data.len() {
+                if buf.len() < data.len() {
                     me.read = State::Ready(data);
                     Err(io::Error::from_raw_os_error(WSAEMSGSIZE as i32))
                 } else {
                     let r = if let Some(addr) = me.read_buf.to_socket_addr() {
-                        buf.write_slice(&data);
-                        Ok(Some(addr))
+                        buf.write(&data).unwrap();
+                        Ok(Some((data.len(), addr)))
                     } else {
                         Err(io::Error::new(io::ErrorKind::Other,
                                            "failed to parse socket address"))